pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = ""
authors = ["nickAnhel <anhimovn1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pydantic-settings = "^2.3.4"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
bcrypt = {extras = ["crypto"], version = "^4.1.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
